// grai-frontend/src/components/chat/__generated__/NewMessage.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: NewMessage
// ====================================================

export interface NewMessage {
  __typename: "Message";
  id: any;
  message: string;
  role: string;
  created_at: any;
}